Update debian defaults in preparation for starting as root.
Previously the debian defaults were only loaded when we started Tor as
the debian-tor user.  Now they are also loaded when started as root.
In addition to logging, pidfile, datadir and runasdaemon we also change
the User and Group options from their NULL default to "debian-tor" so that
Tor correctly setuids/setgids to the role user the package created.

This will allow us to start the daemon as root.


svn:r13685