Build debian's micro-revision into the binary
If we have a debian/micro-revision.i, replace the one in src/or
with our copy so that this will be the revision that ends up in
the binary.  This is an informational only version string, but
it'd be kinda nice if it was (more) accurate nonetheless.

Of course this won't help if people manually patch around but
it's still preferable to claiming we are exactly upstream's source.

If we are building directly out of a git tree, update
debian/micro-revision.i in the clean target.